+++ src/backend/efiling-reviewer-api/src/main/java/ca/bc/gov/open/jag/efilingreviewerapi/extract/DiligenService.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.open.efilingdiligenclient.diligen;
+package ca.bc.gov.open.jag.efilingreviewerapi.extract;
 
 import org.springframework.web.multipart.MultipartFile;
