+++ src/backend/efiling-reviewer-api/src/main/java/ca/bc/gov/open/jag/efilingreviewerapi/extract/DiligenServiceImpl.java
@@ -1,6 +1,8 @@
-package ca.bc.gov.open.efilingdiligenclient.diligen;
+package ca.bc.gov.open.jag.efilingreviewerapi.extract;
 
 import ca.bc.gov.open.efilingdiligenclient.Keys;
+import ca.bc.gov.open.efilingdiligenclient.diligen.DiligenAuthService;
+import ca.bc.gov.open.jag.efilingdiligenclientstarter.DiligenProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.io.ByteArrayResource;
