+++ src/backend/efiling-reviewer-api/src/test/java/ca/bc/gov/open/jag/efilingreviewerapi/extract/diligenServiceImpl/postDocumentTest.java
@@ -1,0 +1,4 @@
+package ca.bc.gov.open.jag.efilingreviewerapi.extract.diligenServiceImpl;
+
+public class postDocumentTest {
+}
