+++ src/backend/libs/efiling-diligen-client-starter/src/main/java/ca/bc/gov/open/jag/efilingdiligenclientstarter/AutoConfiguration.java
@@ -10,7 +10,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
 
 /**
  * Diligen Client configuration
@@ -63,11 +62,6 @@ public class AutoConfiguration {
     public DiligenAuthService diligenAuthService(ApiClient apiClient) {
         AuthenticationApi authenticationApi = new AuthenticationApi(apiClient);
         return new DiligenAuthServiceImpl(authenticationApi);
-    }
-
-    @Bean
-    public DiligenService diligenService(DiligenAuthService diligenAuthService) {
-        return new DiligenServiceImpl(new RestTemplate(), diligenProperties, diligenAuthService);
     }
 
 }
