--- src/backend/libs/efiling-diligen-client/src/test/java/ca/bc/gov/open/efilingdiligenclient/diligen/diligenServiceImpl/postDocumentTest.java
@@ -1,4 +1,0 @@
-package ca.bc.gov.open.efilingdiligenclient.diligen.diligenServiceImpl;
-
-public class postDocumentTest {
-}
